Flask-JWT-Extended is set up

+++ resources/item.py
@@ -1,27 +1,33 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required,
+    get_jwt_claims,
+    jwt_optional,
+    get_jwt_identity,
+    fresh_jwt_required
+)
 from models.item import ItemModel
 
 class Item(Resource):
     parser = reqparse.RequestParser()
     parser.add_argument('price',
-        type=float,
-        required=True,
-        help='This field cannot be left blank!'
-    )
+                        type=float,
+                        required=True,
+                        help='This field cannot be left blank!'
+                        )
     parser.add_argument('store_id',
-        type=int,
-        required=True,
-        help='Every item needs a store id.'
-    )
-
-    @jwt_required()
+                        type=int,
+                        required=True,
+                        help='Every item needs a store id.'
+                        )
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {'message': 'Item not found'}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
@@ -37,12 +43,16 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-        
-        return {'message': 'Item deleted'}
+            return {'message': 'Item deleted'}
+        return {'message': 'Item not found.'}, 404
 
     def put(self, name):
         data = Item.parser.parse_args()
@@ -58,5 +68,13 @@ class Item(Resource):
         return item.json()
 
 class ItemList(Resource):
-	def get(self):
-		return {'items': [item.json() for item in ItemModel.query.all()]}
+    @jwt_optional # The authorization token may be sent or not
+    def get(self):
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200        
+        return {
+                'items': [item['name'] for item in items],
+                'message': 'More data available if you log in.'
+                }, 200
